Fetching BNS records loader dialog added in My BNS screen
