+++ beldex_coin/lib/src/util/types.dart
@@ -135,6 +135,10 @@ typedef CanRequestUnstake = int Function(Pointer<Utf8> masterNodeKey);
 
 typedef BnsSetRecord = int Function(Pointer<Utf8> name);
 
+typedef GetNameToNameHash = Pointer<Utf8> Function(Pointer<Utf8> name);
+
+typedef get_names_to_namehash = Pointer<Utf8> Function(Pointer<Utf8> name);
+
 typedef SubmitStakeUnlock = int Function(Pointer<Utf8> masterNodeKey,
     Pointer<Utf8Box> error, Pointer<PendingTransactionRaw> pendingTransaction);
